Fix slew limiter DSP to accomodate high sample rates and extreme shapes. #221 #220

+++ src/dsp/signal.hpp
@@ -121,7 +121,7 @@ struct ShapedSlewLimiter {
 	float _time;
 	float _shapeExponent;
 	float _inverseShapeExponent;
-	float _last = 0.0;
+	double _last = 0.0;
 
 	ShapedSlewLimiter(float sampleRate = 1000.0f, float milliseconds = 1.0f, float shape = 1.0f) {
 		setParams(sampleRate, milliseconds, shape);
